requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
openpyxl==3.1.2
APScheduler==3.10.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
psutil==5.9.7
scikit-learn==1.3.2
httpx==0.25.2
